apps/api/pyproject.toml
[project]
name = "aegissec-api"
version = "0.1.0"
description = "Local-first API scaffold for authorized security research and validation workflows"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.16.5,<2.0.0",
  "docker>=7.1.0,<8.0.0",
  "fastapi>=0.116.0,<1.0.0",
  "httpx>=0.28.1,<1.0.0",
  "langgraph>=0.6.6,<0.7.0",
  "mcp[cli]>=1.26.0,<2.0.0",
  "PyYAML>=6.0.2,<7.0.0",
  "pydantic-settings>=2.10.1,<3.0.0",
  "sqlmodel>=0.0.24,<0.1.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
]

[dependency-groups]
dev = [
  "black>=25.1.0,<26.0.0",
  "mypy>=1.17.1,<2.0.0",
  "pytest>=8.4.1,<9.0.0",
  "ruff>=0.12.8,<0.13.0",
  "types-PyYAML>=6.0.12.20250822,<7.0.0",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
packages = ["app", "tests"]

[[tool.mypy.overrides]]
module = ["docker", "docker.*"]
ignore_missing_imports = true
